refactor: use userIsTeam and userIsAdmin
