+++ src/Components/SavedSearchAlert/__tests__/SavedSearchAlertModal.jest.tsx
@@ -192,6 +192,7 @@ describe("SavedSearchAlertModal", () => {
         relayEnvironment: {},
         user: {
           email: "me@artsymail.com",
+          roles: ["team"],
         },
       }))
